feat(app): new app style
