+++ app/src/api/styles.json
@@ -6,13 +6,19 @@
 		"backgroundColor": "#FF5B22"
 	},
 	"bg-liane-orange-lighter": {
-		"backgroundColor": "#FF865C"
-	},
-	"bg-liane-orange-light": {
-		"backgroundColor": "#FFB399"
+		"backgroundColor": "#FF7545"
 	},
 	"bg-liane-yellow": {
 		"backgroundColor": "#FFB545"
+	},
+	"bg-liane-royal": {
+		"backgroundColor": "#22278A"
+	},
+	"bg-liane-blue": {
+		"backgroundColor": "#0B79F9"
+	},
+	"bg-liane-pink": {
+		"backgroundColor": "#FF8484"
 	},
 	"bg-orange-light": {
 		"--tw-bg-opacity": 1,
