+++ app/src/screens/LocationWizard.tsx
@@ -1,219 +1,174 @@
-import React, { useCallback, useContext } from "react";
-import { Alert, Image, ImageBackground, View } from "react-native";
-import { RouteProp } from "@react-navigation/native";
-import { StackNavigationProp } from "@react-navigation/stack";
-import * as Location from "expo-location";
-import { AppButton } from "@/components/base/AppButton";
-import { AppText } from "@/components/base/AppText";
+import React, { useContext, useState } from "react";
+import { Alert, Image, ScrollView, View } from "react-native";
 import { AppContext } from "@/components/ContextProvider";
-import { NavigationParamList } from "@/components/Navigation";
-import { tw } from "@/api/tailwind";
+import * as Location from "expo-location";
 import { LocationPermissionLevel } from "@/api";
+import { tw } from "@/api/tailwind";
+import { AppText } from "@/components/base/AppText";
+import { AppButton } from "@/components/base/AppButton";
 
-const image = require("@/assets/images/bg-mountains.jpg");
-const logo = require("@/assets/logo_white.png");
-
-type LocationWizardRouteProp = RouteProp<NavigationParamList, "LocationWizard">;
-
-type LocationWizardNavigationProp = StackNavigationProp<NavigationParamList, "LocationWizard">;
-
-type LocationWizardProps = {
-  route: LocationWizardRouteProp;
-  navigation: LocationWizardNavigationProp;
-};
+const logo = require("@/assets/logo_orange.png");
 
 /**
  * Get the text of the current step.
- * @param step number of the step
+ * @param step current step
  */
 function getWizardText(step: number) {
+  let view;
+
   switch (step) {
     case 0:
-      return (
-        <AppText style={tw("text-center text-lg text-gray-500 mt-20")}>
-          <AppText>
-            Covoiturer à la campagne&nbsp;
+      view = (
+        <ScrollView>
+          <AppText style={tw("text-center text-lg text-white m-3 mt-0")}>
+            Covoiturer à la campagne, en toute liberté.
+            Liane relève vos trajets du quotidien entre des points pré-établis.
           </AppText>
-          <AppText style={tw("font-bold text-gray-800")}>
-            en toute liberté
+          <AppText style={tw("text-center font-bold text-lg text-white m-3")}>
+            Vos trajets sont proposés aux autres utilisateurs de façon 100% anonyme. Vous recevrez ensuite des demandes
+            anonymes des autres utilisateurs, que vous pouvez refuser sans vous justifier.
           </AppText>
+          <AppText style={tw("text-center text-lg text-white m-3")}>
+            Pour cela Liane a besoin d&apos;accéder à vos données de géolocalisation. Vos données de géolocalisation ne sont utilisés dans aucun autre but.
+          </AppText>
+        </ScrollView>
+      );
+      break;
+    case 1:
+      view = (
+        <ScrollView>
+          <AppText style={tw("text-center text-lg text-white m-3 mt-0")}>
+            Pour utiliser Liane a son potentiel maximum, elle doit pouvoir suivre votre déplacement en permanence.
+          </AppText>
+          <AppText style={tw("text-center font-bold text-lg text-white m-3")}>
+            A quel moment voulez-vous autoriser Liane à suivre votre position ?
+          </AppText>
+          <AppText style={tw("text-center italic text-sm text-white m-3")}>
+            Vous pourrez toujours le modifier plus tard.
+          </AppText>
+        </ScrollView>
+      );
+      break;
+    default:
+      view = (
+        <AppText>
+          Lorem ipsum.
         </AppText>
       );
-    case 1:
-      return (
-        <View style={tw("flex items-center")}>
-          <AppText
-            style={tw("text-center text-lg text-white font-bold mt-5")}
-          >
-            je propose mes trajets du quotidien entre des points pré-établits
-          </AppText>
-          <View style={tw("flex items-start bg-gray-100 p-4 opacity-80 mt-5")}>
-            <View style={tw("flex flex-row items-center")}>
-              <AppText style={tw("bg-gray-900 text-white font-bold text-center rounded-full py-1 px-3 m-1 mr-2")}>
-                1
-              </AppText>
-              <AppText style={tw("text-lg font-bold text-gray-600 uppercase")}>
-                je reste anonyme
-              </AppText>
-            </View>
-            <View style={tw("flex flex-row items-center")}>
-              <AppText style={tw("bg-gray-900 text-white font-bold text-center rounded-full py-1 px-3 m-1 mr-2")}>
-                2
-              </AppText>
-              <AppText style={tw("text-lg font-bold text-gray-600 uppercase")}>
-                je reçois des demandes
-              </AppText>
-            </View>
-            <View style={tw("flex flex-row items-center")}>
-              <AppText style={tw("bg-gray-900 text-white font-bold text-center rounded-full py-1 px-3 m-1 mr-2")}>
-                3
-              </AppText>
-              <AppText style={tw("text-lg font-bold text-gray-600 uppercase")}>
-                j&apos;accepte ou pas, sans me justifier
-              </AppText>
-            </View>
-
-          </View>
-        </View>
-      );
-    default:
-    case 2:
-      return (
-        <View style={tw("flex items-center")}>
-          <AppText style={tw("text-center text-lg text-white mt-5")}>
-            c&apos;est un acte citoyen
-          </AppText>
-          <AppText style={tw("text-center text-lg text-gray-600 mt-5")}>
-            j&apos;autorise liane à lire ma position GPS
-            pour permettre le partage de mes trajets
-          </AppText>
-          <View style={tw("flex items-center bg-gray-100 p-2 opacity-80 mt-5")}>
-            <AppText style={tw("text-lg font-bold text-gray-800 uppercase")}>c&apos;est le seul usage</AppText>
-          </View>
-        </View>
-      );
   }
+
+  return view;
 }
 
 /**
- * Get the text of the button of the current step.
- * @param step number of the step
+ * Alter template.
  */
-function getButtonText(step: number) {
-  switch (step) {
-    case 0:
-      return "J'y vais";
-    case 1:
-      return "C'est ma liberté";
-    default:
-    case 2:
-      return "Quand l'application est active";
-    case 3:
-      return "Toujours";
-    case 4:
-      return "Jamais";
-  }
+function alert(message: string, callback: Function) {
+  Alert.alert(
+    "Information",
+    message,
+    [
+      {
+        text: "Annuler",
+        style: "cancel"
+      },
+      {
+        text: "Continuer",
+        onPress: async () => callback()
+      }
+    ],
+    { cancelable: true }
+  );
 }
 
-const LocationWizard = ({ route, navigation }: LocationWizardProps) => {
-  const { setLocationPermissionLevel } = useContext(AppContext);
-  const { step = 0 } = route.params;
+/**
+ * React component.
+ */
+const LocationWizard2 = () => {
+  const { setLocationPermissionLevel, authUser } = useContext(AppContext);
+  console.log(authUser);
+  const [step, setStep] = useState(authUser ? 1 : 0);
+  console.log(step);
+  const [optionalText, setOptionalText] = useState("");
 
-  // Advance one step
-  const next = useCallback(async () => {
-    navigation.push("LocationWizard", { step: step + 1 });
-  }, [step]);
+  // Go to next step
+  const next = () => {
+    if (step < 1) {
+      setStep(step + 1);
+    }
+  };
 
-  // Request to give no permissions
-  const requestNoLocPerm = useCallback(async () => {
-    Alert.alert(
-      "Localisation",
-      "Liane a besoin de tracer votre localisation GPS afin de fonctionner correctement. Sans ça, l'application "
-        + "n'est pas utilisable. Vous pourrez toujours l'activer plus tard.",
-      [
-        {
-          text: "Annuler",
-          style: "cancel"
-        },
-        {
-          text: "Continuer",
-          onPress: () => next
+  // Ask for foreground tracking permission
+  const requestForegroundLocPerm = async () => {
+    alert("Liane collecte la position GPS de votre téléphone pour permettre l'enregistrement anonyme de vos trajets lorsque l'application est en cours d'utilisation.", async () => {
+      const permission = await Location.requestForegroundPermissionsAsync();
+      console.log(`New foreground permission status : ${permission.status}, ${permission.canAskAgain}`);
+
+      if (permission) {
+        if (permission.status === "granted") {
+          setLocationPermissionLevel(LocationPermissionLevel.ACTIVE);
+        } else if (permission.canAskAgain) {
+          setOptionalText("Vous avez empêché Liane de re-demander cette permission, rendez-vous dans les paramètres pour la modifier.");
         }
-      ],
-      {
-        cancelable: true
       }
-    );
-  }, [next]);
+    }); // Important alert in order to get validated by Google
+  };
 
-  // Request to give foreground permissions
-  const requestForegroundLocPerm = useCallback(async () => {
-    const permission = await Location.requestForegroundPermissionsAsync();
+  // Ask for background tracking permission
+  const requestBackgroundLocPerm = async () => {
+    alert("Liane collecte la position GPS de votre téléphone pour permettre l'enregistrement anonyme de vos trajets même lorsque l'application est fermée ou n'est pas en cours d'utilisation.", async () => {
+      const permissionForeground = await Location.requestForegroundPermissionsAsync(); // Forced to ask for foreground perms first
 
-    if (permission.status === "granted") {
-      setLocationPermissionLevel(LocationPermissionLevel.ACTIVE);
-      await next();
-    }
+      console.log(`New foreground permission status : ${permissionForeground.status}, ${permissionForeground.canAskAgain}`);
+      if (permissionForeground) {
+        if (permissionForeground.status === "granted") {
+          const permissionBackground = await Location.requestBackgroundPermissionsAsync();
+          console.log(`New background permission status : ${permissionBackground.status}, ${permissionBackground.canAskAgain}`);
 
-  }, [setLocationPermissionLevel, next]);
+          if (permissionBackground) {
+            if (permissionBackground.status === "granted") {
+              setLocationPermissionLevel(LocationPermissionLevel.ALWAYS);
+            } else if (permissionBackground.canAskAgain) {
+              setLocationPermissionLevel(LocationPermissionLevel.ACTIVE);
+              setOptionalText("Vous avez empêché Liane de re-demander cette permission, rendez-vous dans les paramètres pour la modifier.");
+            } else {
+              setLocationPermissionLevel(LocationPermissionLevel.ACTIVE);
+            }
+          }
+        } else if (!permissionForeground.canAskAgain) {
+          setOptionalText("Vous avez empêché Liane de re-demander cette permission, rendez-vous dans les paramètres pour la modifier.");
+        }
+      }
+    });
+  };
 
-  // Request to give background permissions
-  const requestBackgroundLocPerm = useCallback(async () => {
-    const permission = await Location.requestBackgroundPermissionsAsync();
-
-    if (permission.status === "granted") {
-      setLocationPermissionLevel(LocationPermissionLevel.ALWAYS);
-      await next();
-    }
-
-  }, [setLocationPermissionLevel, next]);
+  // Ask for no tracking (for now)
+  const requestNoLocPerm = async () => {
+    setLocationPermissionLevel(LocationPermissionLevel.NOT_NOW);
+  };
 
   return (
-    <ImageBackground source={image} style={tw("h-full min-h-full")} resizeMode="cover">
-      <View style={tw("h-20 items-center mx-20 mt-32")}>
+    <View style={tw("h-full min-h-full bg-liane-yellow content-center")}>
+      <View style={tw("h-10 items-center my-20")}>
         <Image
           style={tw("flex-1 w-64")}
           source={logo}
           resizeMode="contain"
         />
       </View>
-      <View>
-        {getWizardText(step)}
-        {step < 2
-        && (
-        <AppButton
-          buttonStyle={tw("bg-orange-light text-white font-bold py-2 px-1 mr-4 ml-4 my-3 rounded")}
-          titleStyle={tw("text-xl text-white font-bold ")}
-          onPress={next} // Popup for foreground permission
-          title={getButtonText(step)}
-        />
+      { getWizardText(step) }
+      { optionalText !== "" && (<AppText style={tw("text-center text-sm text-red-600 m-2")}>{ optionalText }</AppText>) }
+      { step === 0
+        ? (<View><AppButton buttonStyle={tw("bg-liane-orange rounded-full m-10")} onPress={next} title="Continuer" /></View>)
+        : (
+          <View style={tw("my-10")}>
+            <AppButton buttonStyle={tw("bg-liane-orange rounded-full mt-5 ml-10 mr-10")} onPress={requestBackgroundLocPerm} title="Toujours" />
+            <AppButton buttonStyle={tw("bg-liane-orange-lighter rounded-full mt-5 ml-10 mr-10")} onPress={requestForegroundLocPerm} title="Pendant l'utilisation" />
+            <AppButton buttonStyle={tw("bg-liane-orange-lighter rounded-full mt-5 ml-10 mr-10")} onPress={requestNoLocPerm} title="Jamais (pour le moment)" />
+          </View>
         )}
-        {step >= 2
-          && (
-          <>
-            <AppButton
-              buttonStyle={tw("bg-orange-light text-white font-bold py-2 px-1 mr-4 ml-4 my-3 rounded")}
-              titleStyle={tw("text-xl text-white font-bold ")}
-              onPress={requestBackgroundLocPerm} // Popup for background permission
-              title="Tout le temps"
-            />
-            <AppButton
-              buttonStyle={tw("bg-orange-light text-white font-bold py-2 px-1 mr-4 ml-4 my-3 rounded")}
-              titleStyle={tw("text-xl text-white font-bold ")}
-              onPress={requestForegroundLocPerm} // Popup for background permission
-              title="Quand l'application est active"
-            />
-            <AppButton
-              buttonStyle={tw("bg-orange-light text-white font-bold py-2 px-1 mr-4 ml-4 my-3 rounded")}
-              titleStyle={tw("text-xl text-white font-bold ")}
-              onPress={requestNoLocPerm} // Popup saying you cannot use Liane ?
-              title="Jamais"
-            />
-          </>
-          )}
-      </View>
-    </ImageBackground>
+    </View>
   );
 };
 
-export default LocationWizard;
+export default LocationWizard2;
